requirements.txt
asgiref==3.2.10
Django==3.1.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
pytz==2020.1
sqlparse==0.4.1
